test(abstract-eth): move tests of eth-like coins to abstract-eth TICKET: COIN-1322
