+++ modules/abstract-eth/src/index.ts
@@ -2,3 +2,4 @@ export * from './abstractEthLikeCoin';
 export * from './ethLikeToken';
 export * from './lib';
 export * from './abstractEthLikeNewCoins';
+export * from '../test';
