+++ modules/abstract-eth/tsconfig.json
@@ -7,7 +7,7 @@
     "esModuleInterop": true,
     "typeRoots": ["../../types", "./node_modules/@types", "../../node_modules/@types"]
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "test/**/*"],
   "exclude": ["node_modules"],
   "references": [
     {
@@ -15,6 +15,12 @@
     },
     {
       "path": "../statics"
+    },
+    {
+      "path": "../sdk-api"
+    },
+    {
+      "path": "../sdk-test"
     }
   ]
 }
