+++ modules/sdk-coin-arbeth/test/resources.ts
@@ -37,6 +37,8 @@ export const PRIVATE_KEY_2 =
 export const PUBLIC_KEY_2 =
   'xpub661MyMwAqRbcFXrZT6GpipqTPdFdhYX5DFDGv5ws1iW8ViCWSTK3zVhVntpiKDjPqXrqP1hhQscJZhVpbmZt9qygxgPdmN9RAZVKqovPfLw';
 
+export const ACCOUNT_3 = '0xb1938215967408fff7c59c77ae5e5283b55c8e26';
+
 export const SEND_FUNDS_DATA =
   '0x391252150000000000000000000000007325a3f7d4f9e86ae62cf742426078c3755730d5000000000000000000000000000000000000000000000000016345785d8a000000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000005ec6ab34000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041d8decc5aa3e56cba81cb5405cdd187a01885fd118ca3add06fc10c471d86b94459fe0c0b7a906e2147f64d948616236241288d8bf775873692b9963803ac608c1c00000000000000000000000000000000000000000000000000000000000000';
 
@@ -79,6 +81,8 @@ export const BASE_ADDRESS = '0xbc42187c73d2f41a5551be6d138298f4e9f2e2c6';
 
 export const FEE_ADDRESS = '0x5fc4bc0f18c4243ee4ce757d712bb2164176cb51';
 
+export const DEPLOYED_ADDRESS = '0xd9d189aa6907a4fafd3c66e8cc178011b1f443f5';
+
 export const UNSIGNED_ADDRESS_INIT =
   '0xf887010a8203e89437996e762fa8b671869740c79eb33f625b3bf92a80b86413b2f75c000000000000000000000000bc42187c73d2f41a5551be6d138298f4e9f2e2c60000000000000000000000005fc4bc0f18c4243ee4ce757d712bb2164176cb510000000000000000000000000000000000000000000000000000000000000001830cddff8080';
 
@@ -87,3 +91,21 @@ export const SIGNED_ADDRESS_INIT =
 
 export const TX_BROADCAST =
   '0xf9012d018502540be4008367c28094809ee567e413543af1caebcdb247f6a67eafc8dd80b8c47117f3fa000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003000000000000000000000000e6c43626f11312de29b0011fa9da71ea3bba0e9f00000000000000000000000078caeb4527170e52f54d936e4eef6f83250e01bb000000000000000000000000b1938215967408fff7c59c77ae5e5283b55c8e26830cde00a0135b3d12f8b34d3658d0cb81b42f1bca0847733f9ec904cab1d3adfe13c8eb23a014d8dd111c95fa547602c4828314d27dbb49b4cd8e3201665ae11146ef9e7f11';
+
+export const SIGNED_INIT_TX_ID = '0xdc7ebdb8aa3c190888825fb38a29ff0530247a436df0204577b95de96bf6baef';
+
+export const CHAIN_ID = '421614';
+
+export const FINAL_V = '0x0cddff';
+
+export const NETWORK_TOKEN_IDENTIFIER = 'tarbeth:link';
+
+export const COIN = 'tarbeth';
+
+export const EXPECTED_NEW_TX_ID = '0x2753999553d3602e65fc272738608515f154f83adf37c576118a87f5bc9309f9';
+
+export const CHAIN_FULL_NAME = 'Arbitrum Ethereum (L2 Chain)';
+
+export const BASE_URL = 'https://api-sepolia.arbiscan.io';
+
+export const BACKUP_KEY_ADDRESS = '0x4f2c4830cc37f2785c646f89ded8a919219fa0e9';
