+++ modules/sdk-coin-bsc/test/unit/bscToken.ts
@@ -1,32 +1,15 @@
 import 'should';
-
-import { TestBitGo, TestBitGoAPI } from '@bitgo/sdk-test';
 import { BscToken } from '../../src';
-import { BitGoAPI } from '@bitgo/sdk-api';
+import { runTokenTestInitialization } from '@bitgo/abstract-eth';
+import * as testData from '../resources/bsc';
 
-describe('Bsc Token:', function () {
-  let bitgo: TestBitGoAPI;
-  let bscTokenCoin;
-  const tokenName = 'tbsc:busd';
+describe('Bsc Token Tests', () => {
+  const coinName = 'Bsc';
+  const tokenNetworkName = 'Test Binance USD Token';
 
-  before(function () {
-    bitgo = TestBitGo.decorate(BitGoAPI, { env: 'test' });
-    BscToken.createTokenConstructors().forEach(({ name, coinConstructor }) => {
-      bitgo.safeRegister(name, coinConstructor);
+  describe('Bsc tokens in test env:', () => {
+    it('Bsc run token tests', () => {
+      runTokenTestInitialization(BscToken, coinName, tokenNetworkName, testData);
     });
-    bitgo.initializeTestVars();
-    bscTokenCoin = bitgo.coin(tokenName);
-  });
-
-  it('should return constants', function () {
-    bscTokenCoin.getChain().should.equal('tbsc:busd');
-    bscTokenCoin.getBaseChain().should.equal('tbsc');
-    bscTokenCoin.getFullName().should.equal('Bsc Token');
-    bscTokenCoin.getBaseFactor().should.equal(1e18);
-    bscTokenCoin.type.should.equal(tokenName);
-    bscTokenCoin.name.should.equal('Test Binance USD Token');
-    bscTokenCoin.coin.should.equal('tbsc');
-    bscTokenCoin.network.should.equal('Testnet');
-    bscTokenCoin.decimalPlaces.should.equal(18);
   });
 });
