+++ modules/sdk-coin-opeth/test/resources.ts
@@ -37,6 +37,8 @@ export const PRIVATE_KEY_2 =
 export const PUBLIC_KEY_2 =
   'xpub661MyMwAqRbcFXrZT6GpipqTPdFdhYX5DFDGv5ws1iW8ViCWSTK3zVhVntpiKDjPqXrqP1hhQscJZhVpbmZt9qygxgPdmN9RAZVKqovPfLw';
 
+export const ACCOUNT_3 = '0xb1938215967408fff7c59c77ae5e5283b55c8e26';
+
 export const SEND_FUNDS_DATA =
   '0x391252150000000000000000000000007325a3f7d4f9e86ae62cf742426078c3755730d5000000000000000000000000000000000000000000000000016345785d8a000000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000005ec6ab34000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000e00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004106935b57df83e029901bb12028bcd86feff0724699ae71bee552dd5ed99766b2794d4cc6dfd8f94949f21fea3188c2cd502c89bd0adb939bcf9dd7e79e8f7aca1b00000000000000000000000000000000000000000000000000000000000000';
 
@@ -79,6 +81,8 @@ export const BASE_ADDRESS = '0xbc42187c73d2f41a5551be6d138298f4e9f2e2c6';
 
 export const FEE_ADDRESS = '0x5fc4bc0f18c4243ee4ce757d712bb2164176cb51';
 
+export const DEPLOYED_ADDRESS = '0xd9d189aa6907a4fafd3c66e8cc178011b1f443f5';
+
 export const UNSIGNED_ADDRESS_INIT =
   '0xf888010a8203e89437996e762fa8b671869740c79eb33f625b3bf92a80b86413b2f75c000000000000000000000000bc42187c73d2f41a5551be6d138298f4e9f2e2c60000000000000000000000005fc4bc0f18c4243ee4ce757d712bb2164176cb5100000000000000000000000000000000000000000000000000000000000000018401546fdb8080';
 
@@ -87,3 +91,17 @@ export const SIGNED_ADDRESS_INIT =
 
 export const TX_BROADCAST =
   '0xf9012e018502540be4008367c28094809ee567e413543af1caebcdb247f6a67eafc8dd80b8c47117f3fa000000000000000000000000000000000000000000000000000000000000004000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000003000000000000000000000000e6c43626f11312de29b0011fa9da71ea3bba0e9f00000000000000000000000078caeb4527170e52f54d936e4eef6f83250e01bb000000000000000000000000b1938215967408fff7c59c77ae5e5283b55c8e268401546fdba0d5fb980bc38c3276807db5a2d54307f1bd29d157ecb16b691f49abda0a314a55a0043253e485a3d33d1d63db2d77cf0caac8b0d1d85dc7dbcec0595ac764a7608f';
+
+export const CHAIN_ID = '11155420';
+
+export const NETWORK_TOKEN_IDENTIFIER = 'topeth:terc18dp';
+
+export const COIN = 'topeth';
+
+export const EXPECTED_NEW_TX_ID = '0x4f9ac32512aeef15e2b3e4bd79643c6f77f93b462f20f1f3d3e48d5905d12add';
+
+export const FINAL_V = '0x01546fdb';
+
+export const CHAIN_FULL_NAME = 'Optimism Ethereum (L2 Chain)';
+
+export const BASE_URL = 'https://api-sepolia-optimistic.etherscan.io';
