+++ modules/sdk-coin-opeth/test/unit/transferBuilder.ts
@@ -1,7 +1,6 @@
 import should from 'should';
 import { coins, EthereumNetwork as EthLikeNetwork } from '@bitgo/statics';
 import { KeyPair, TransferBuilder } from '../../src';
-
 import * as testData from '../resources';
 
 describe('Opeth send multi sig builder', function () {
@@ -10,11 +9,13 @@ describe('Opeth send multi sig builder', function () {
     'xprv9s21ZrQH143K3D8TXfvAJgHVfTEeQNW5Ys9wZtnUZkqPzFzSjbEJrWC1vZ4GnXCvR7rQL2UFX3RSuYeU9MrERm1XBvACow7c36vnz5iYyj2';
   const key = new KeyPair({ prv: xprv }).getKeys().prv as string;
   const amount = '100000000000000000'; // equivalent to 0.1 ether
-  const coin = coins.get('topeth') as unknown as EthLikeNetwork;
+  const coinName = testData.COIN;
+  const networkTokenIdentifier = testData.NETWORK_TOKEN_IDENTIFIER;
+  const coin = coins.get(coinName) as unknown as EthLikeNetwork;
   describe('should build', () => {
     it('native coin transfer should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('topeth')
+        .coin(coinName)
         .expirationTime(1590078260)
         .amount(amount)
         .to(toAddress)
@@ -27,7 +28,7 @@ describe('Opeth send multi sig builder', function () {
 
     it('native coin transfer with sequenceId zero should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('topeth')
+        .coin(coinName)
         .expirationTime(1590078260)
         .amount(amount)
         .to(toAddress)
@@ -40,7 +41,7 @@ describe('Opeth send multi sig builder', function () {
 
     it('native coin transfer with amount 0 should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('topeth')
+        .coin(coinName)
         .expirationTime(1590078260)
         .amount('0')
         .to(toAddress)
@@ -53,7 +54,7 @@ describe('Opeth send multi sig builder', function () {
 
     it('ERC20 token transfer should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('topeth:terc18dp')
+        .coin(networkTokenIdentifier)
         .expirationTime(1590078260)
         .amount(amount)
         .to(toAddress)
@@ -65,7 +66,7 @@ describe('Opeth send multi sig builder', function () {
 
     it('erc20 transfer should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('topeth:terc18dp')
+        .coin(networkTokenIdentifier)
         .expirationTime(1590078260)
         .amount(amount)
         .to(toAddress)
@@ -77,7 +78,7 @@ describe('Opeth send multi sig builder', function () {
 
     it('erc20 transfer with amount 0 should succeed', async () => {
       const builder = new TransferBuilder()
-        .coin('topeth:terc18dp')
+        .coin(networkTokenIdentifier)
         .expirationTime(1590078260)
         .amount('0')
         .to(toAddress)
@@ -100,7 +101,7 @@ describe('Opeth send multi sig builder', function () {
 
     it('should build from a non signed serialized data', () => {
       const builder = new TransferBuilder(testData.SEND_FUNDS_NO_KEY_DATA);
-      builder.coin('topeth').key(key);
+      builder.coin(coinName).key(key);
       const result = builder.signAndBuild(`${coin.chainId}`);
       should.equal(result, testData.SEND_FUNDS_DATA);
     });
