+++ modules/sdk-coin-polygon/test/resources.ts
@@ -75,3 +75,18 @@ export const RECOVERY_WALLET_BYTE_CODE =
 
 export const SEND_TX_BROADCAST_LEGACY =
   '0xf901cd02843b9aca0083b8a1a0948f977e912ef500548a0c3be6ddde9899f1199b8180b901643912521500000000000000000000000019645032c7f1533395d44a629462e751084d3e4c000000000000000000000000000000000000000000000000000000003b9aca0000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000005ec67e28000000000000000000000000000000000000000000000000000000000000000500000000000000000000000000000000000000000000000000000000000000e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000041caf7a8fc689017e36f25e5d53982b19457499198957176feb74d291e333537ef46528e725032da1bb9b8dbf8f3f5ca6f14b4cc6ce79ff94f7dba4f217b29de971b0000000000000000000000000000000000000000000000000000000000000083027127a0c6beaec31e6d55b69a2fa0b6f0812f90721709d2bf5eaca0196e9fa45b4890ffa07c7283c70b13ca41e1779a3f52aaf72f5193dd900f366752d7c6dadba43067e9';
+
+export const DEPLOYED_ADDRESS = '0xde4133877caa961ff30caf3373c5a2f9e9cd31b2';
+
+export const COIN = 'tpolygon';
+
+export const EXPECTED_NEW_TX_ID = '0x9c955e671f05f8e4e909757675f535bdbcdb95cc66fbb4e897146dec20383303';
+
+export const FINAL_V = '0x027127';
+
+export const NETWORK_TOKEN_IDENTIFIER = 'tpolygon:derc20';
+
+export const BASE_URL = 'https://api-amoy.polygonscan.com';
+
+export const SEND_TOKEN_TX_BROADCAST =
+  '0xf901ad02843b9aca0083b8a1a094df7decb1baa8f529f0c8982cbb4be5035719529980b901440dcd7a6c00000000000000000000000072c2c8e08bf91d755cd7d26b49a2ee3dc99de1b90000000000000000000000000000000000000000000000000000000000000064000000000000000000000000fe4f5145f6e09952a5ba9e956ed0c25e3fa4c7f1000000000000000000000000000000000000000000000000000000005ec67e28000000000000000000000000000000000000000000000000000000000000000500000000000000000000000000000000000000000000000000000000000000c00000000000000000000000000000000000000000000000000000000000000041e083101304ea640e3ad0f7a627a0f788c6a100e64a9454934e67cfc91fd691c76153550ec2aa357791fc0ed54471938d8ce7c501a261027797e54b18bda5a4301c0000000000000000000000000000000000000000000000000000000000000083027127a0599be8e8ff280a7fc04b45cc735ebc21bf1934c6da1acc9777e09570014f0ed5a0121a6ecd26c0ac264b331fa97c8eac0324581b76c151b4be742fbe313b3dde57';
