+++ modules/sdk-coin-zketh/test/resources.ts
@@ -62,3 +62,16 @@ export const SEND_TOKEN_AMOUNT_ZERO_DATA =
 
 export const SEND_FUNDS_NO_KEY_DATA =
   '0x391252150000000000000000000000007325a3f7d4f9e86ae62cf742426078c3755730d5000000000000000000000000000000000000000000000000016345785d8a000000000000000000000000000000000000000000000000000000000000000000c0000000000000000000000000000000000000000000000000000000005ec6ab34000000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000e000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000';
+
+export const SEND_TOKEN_TX_BROADCAST =
+  '0xf901ac02843b9aca0083b8a1a094df7decb1baa8f529f0c8982cbb4be5035719529980b901440dcd7a6c00000000000000000000000072c2c8e08bf91d755cd7d26b49a2ee3dc99de1b90000000000000000000000000000000000000000000000000000000000000064000000000000000000000000cccb29bac5ad81290383643c6fb38130cda9d881000000000000000000000000000000000000000000000000000000005ec67e28000000000000000000000000000000000000000000000000000000000000000500000000000000000000000000000000000000000000000000000000000000c000000000000000000000000000000000000000000000000000000000000000419c5db173f0b5abcc1fe40a7a455e78feeca5128e25619a7d1f7ae3bd111fd0af5426c1148c42d410ba73f27e698d3cf458234afdfc0df383f5e0486da667bb491b0000000000000000000000000000000000000000000000000000000000000082027ba069e9a05a73f246f012f06dc44b5d736f2151b80ce6cbed642de5a568ada21698a01423f7fcbff7bdde497ddd4c69142ba9082ff69a9262cb37cd0ff5f2685ec427';
+
+export const CHAIN_ID = '300';
+
+export const COIN = 'tzketh';
+
+export const NETWORK_TOKEN_IDENTIFIER = 'tzketh:link';
+
+export const CHAIN_FULL_NAME = 'zkSync Ethereum';
+
+export const BASE_URL = 'https://block-explorer-api.sepolia.zksync.dev';
